Fix macOS static library build.
On Linux, we use GNU ar to build the final libnewrelic.a, since it has
support for a DSL to perform all sorts of operations, including merging
static libraries.

On macOS, using GNU ar is a really good way to get inscrutable error
messages around multi-architecture support, as the system ar has a bunch
of magic to handle that. However, macOS ar, being derived from BSD,
doesn't have the fancy GNU DSL.

macOS ships with a `libtool` that is supposed to subsume the
functionality provided by `ar` and `ranlib` (but is not to be confused
with autotools' libtool). Therefore, on macOS, we'll use that to combine
the libraries instead.